+++ data/src/main/java/com/kyj/data/remote/datasourceimpl/CoronaCenterRemoteDataSourceImpl.kt
@@ -1,7 +1,7 @@
-package com.kyj.data.datasourceimpl
+package com.kyj.data.remote.datasourceimpl
 
 import com.kyj.data.common.util.safeApiCall
-import com.kyj.data.datasource.CoronaCenterRemoteDataSource
+import com.kyj.data.remote.datasource.CoronaCenterRemoteDataSource
 import com.kyj.data.remote.api.CoronaCenterApiService
 import com.kyj.data.remote.response.CoronaCenterResponse
 import com.kyj.domain.util.NetworkResult
@@ -10,7 +10,7 @@ import javax.inject.Inject
 class CoronaCenterRemoteDataSourceImpl @Inject constructor(
     private val coronaCenterApiService: CoronaCenterApiService,
 ) : CoronaCenterRemoteDataSource {
-    override suspend fun getCoronaCenters(page: Int): NetworkResult<CoronaCenterResponse> {
-        return safeApiCall { coronaCenterApiService.getCoronaCenters(page) }
+    override suspend fun downloadCoronaCenters(page: Int): NetworkResult<CoronaCenterResponse> {
+        return safeApiCall { coronaCenterApiService.downloadCoronaCenters(page) }
     }
 }
