+++ domain/src/main/java/com/kyj/domain/usecase/GetCoronaCentersUseCase.kt
@@ -1,12 +1,12 @@
 package com.kyj.domain.usecase
 
-import com.kyj.domain.model.CoronaCentersInfo
+import com.kyj.domain.model.CoronaCenter
 import com.kyj.domain.repository.CoronaCenterRepository
-import com.kyj.domain.util.NetworkResult
+import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class GetCoronaCentersUseCase @Inject constructor(
     private val coronaCenterRepository: CoronaCenterRepository,
 ) {
-    suspend operator fun invoke(page: Int): NetworkResult<CoronaCentersInfo> = coronaCenterRepository.getCoronaCenters(page)
+    operator fun invoke(): Flow<List<CoronaCenter>> = coronaCenterRepository.getCoronaCenters()
 }
